+++ src/test/java/org/apache/datasketches/kll/KllDoublesSketchIteratorTest.java
@@ -54,7 +54,7 @@ public class KllDoublesSketchIteratorTest {
       int weight = 0;
       while (it.next()) {
         count++;
-        weight += it.getWeight();
+        weight += (int)it.getWeight();
       }
       Assert.assertEquals(count, sketch.getNumRetained());
       Assert.assertEquals(weight, n);
